Changed Float to Double Added stuff for publication on maven

+++ binding/src/main/kotlin/ch/viseon/threejs/declarations/animation/AnimationAction.kt
@@ -1,4 +1,4 @@
-//Generated date 22.11.2018
+//Generated date 23.11.2018
 @file:JsModule("three")
 package ch.viseon.threejs.declarations.animation
 
@@ -22,7 +22,7 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* The looping mode (can be changed with [page:.setLoop setLoop]). Default is [page:Animation THREE.LoopRepeat] (with an infinite number of [page:.repetitions repetitions]) Must be one of these constants: [page:Animation THREE.LoopOnce] - playing the clip once, [page:Animation THREE.LoopRepeat] - playing the clip with the choosen number of **repetitions**, each time jumping from the end of the clip directly to its beginning, [page:Animation THREE.LoopPingPong] - playing the clip with the choosen number of **repetitions**, alternately playing forward and backward.
 	*/
-	open var loop: Float  = definedExternally
+	open var loop: Double  = definedExternally
 
 
 	/**
@@ -34,25 +34,25 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* The number of repetitions of the performed [page:AnimationClip] over the course of this action. Can be set via [page:.setLoop setLoop]. Default is **Infinity**. Setting this number has no effect, if the [page:.loop loop mode] is set to [page:Animation THREE.LoopOnce].
 	*/
-	open var repetitions: Float  = definedExternally
+	open var repetitions: Double  = definedExternally
 
 
 	/**
 	* The local time of this action (in seconds, starting with 0). The value gets clamped or wrapped to 0...clip.duration (according to the loop state). It can be scaled relativly to the global mixer time by changing [page:.timeScale timeScale] (using [page:.setEffectiveTimeScale setEffectiveTimeScale] or [page:.setDuration setDuration]).
 	*/
-	open var time: Float  = definedExternally
+	open var time: Double  = definedExternally
 
 
 	/**
 	* Scaling factor for the [page:.time time]. A value of 0 causes the animation to pause. Negative values cause the animation to play backwards. Default is 1. Properties/methods concerning **timeScale** (respectively **time**) are: [page:.getEffectiveTimeScale getEffectiveTimeScale], [page:.halt halt], [page:.paused paused], [page:.setDuration setDuration], [page:.setEffectiveTimeScale setEffectiveTimeScale], [page:.stopWarping stopWarping], [page:.syncWith syncWith], [page:.warp warp].
 	*/
-	open var timeScale: Float  = definedExternally
+	open var timeScale: Double  = definedExternally
 
 
 	/**
 	* The degree of influence of this action (in the interval [0, 1]). Values between 0 (no impact) and 1 (full impact) can be used to blend between several actions. Default is 1. Properties/methods concerning **weight** are: [page:.crossFadeFrom crossFadeFrom], [page:.crossFadeTo crossFadeTo], [page:.enabled enabled], [page:.fadeIn fadeIn], [page:.fadeOut fadeOut], [page:.getEffectiveWeight getEffectiveWeight], [page:.setEffectiveWeight setEffectiveWeight], [page:.stopFading stopFading].
 	*/
-	open var weight: Float  = definedExternally
+	open var weight: Double  = definedExternally
 
 
 	/**
@@ -70,37 +70,37 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* Causes this action to [page:.fadeIn fade in], fading out another action simultaneously, within the passed time interval. This method can be chained. If warpBoolean is true, additional [page:.warp warping] (gradually changes of the time scales) will be applied. Note: Like with **fadeIn** / **fadeOut**, the fading starts/ends with a weight of 1.
 	*/
-	open fun crossFadeFrom(fadeOutAction: ch.viseon.threejs.declarations.animation.AnimationAction = definedExternally, durationInSeconds: Float = definedExternally, warpBoolean: Boolean = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun crossFadeFrom(fadeOutAction: ch.viseon.threejs.declarations.animation.AnimationAction = definedExternally, durationInSeconds: Double = definedExternally, warpBoolean: Boolean = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Causes this action to [page:.fadeOut fade out], fading in another action simultaneously, within the passed time interval. This method can be chained. If warpBoolean is true, additional [page:.warp warping] (gradually changes of the time scales) will be applied. Note: Like with **fadeIn** / **fadeOut**, the fading starts/ends with a weight of 1.
 	*/
-	open fun crossFadeTo(fadeInAction: ch.viseon.threejs.declarations.animation.AnimationAction = definedExternally, durationInSeconds: Float = definedExternally, warpBoolean: Boolean = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun crossFadeTo(fadeInAction: ch.viseon.threejs.declarations.animation.AnimationAction = definedExternally, durationInSeconds: Double = definedExternally, warpBoolean: Boolean = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Increases the [page:.weight weight] of this action gradually from 0 to 1, within the passed time interval. This method can be chained.
 	*/
-	open fun fadeIn(durationInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun fadeIn(durationInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Decreases the [page:.weight weight] of this action gradually from 1 to 0, within the passed time interval. This method can be chained.
 	*/
-	open fun fadeOut(durationInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun fadeOut(durationInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Returns the effective time scale (considering the current states of warping and [page:.paused paused]).
 	*/
-	open fun getEffectiveTimeScale() : Float
+	open fun getEffectiveTimeScale() : Double
 
 
 	/**
 	* Returns the effective weight (considering the current states of fading and [page:.enabled enabled]).
 	*/
-	open fun getEffectiveWeight() : Float
+	open fun getEffectiveWeight() : Double
 
 
 	/**
@@ -124,7 +124,7 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* Decelerates this animation's speed to 0 by decreasing [page:.timeScale timeScale] gradually (starting from its current value), within the passed time interval. This method can be chained.
 	*/
-	open fun halt(durationInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun halt(durationInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
@@ -154,31 +154,31 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* Sets the duration for a single loop of this action (by adjusting [page:.timeScale timeScale] and stopping any scheduled warping). This method can be chained.
 	*/
-	open fun setDuration(durationInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun setDuration(durationInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Sets the [page:.timeScale timeScale] and stops any scheduled warping. This method can be chained. If [page:.paused paused] is false, the effective time scale (an internal property) will also be set to this value; otherwise the effective time scale (directly affecting the animation at this moment) will be set to 0. Note: .**paused** will not be switched to **true** automatically, if .**timeScale** is set to 0 by this method.
 	*/
-	open fun setEffectiveTimeScale(timeScale: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun setEffectiveTimeScale(timeScale: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Sets the [page:.weight weight] and stops any scheduled fading. This method can be chained. If [page:.enabled enabled] is true, the effective weight (an internal property) will also be set to this value; otherwise the effective weight (directly affecting the animation at this moment) will be set to 0. Note: .**enabled** will not be switched to **false** automatically, if .**weight** is set to 0 by this method.
 	*/
-	open fun setEffectiveWeight(weight: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun setEffectiveWeight(weight: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Sets the [page:.loop loop mode] and the number of [page:.repetitions repetitions]. This method can be chained.
 	*/
-	open fun setLoop(loopMode: Float = definedExternally, repetitions: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun setLoop(loopMode: Double = definedExternally, repetitions: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
 	* Defines the time for a delayed start (usually passed as [page:AnimationMixer.time] + deltaTimeInSeconds). This method can be chained. Note: The animation will only start at the given time, if .**startAt** is chained with [page:.play play], or if the action has already been activated in the mixer (by a previous call of .**play**, without stopping or resetting it in the meantime).
 	*/
-	open fun startAt(startTimeInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun startAt(startTimeInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 
 
 	/**
@@ -208,5 +208,5 @@ open external class AnimationAction(mixer: ch.viseon.threejs.declarations.animat
 	/**
 	* Changes the playback speed, within the passed time interval, by modifying [page:.timeScale timeScale] gradually from **startTimeScale** to **endTimeScale**. This method can be chained.
 	*/
-	open fun warp(startTimeScale: Float = definedExternally, endTimeScale: Float = definedExternally, durationInSeconds: Float = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
+	open fun warp(startTimeScale: Double = definedExternally, endTimeScale: Double = definedExternally, durationInSeconds: Double = definedExternally) : ch.viseon.threejs.declarations.animation.AnimationAction
 }
